App Icon Updated Re-organized Files myProfile Progress Home + Adapter + FoodDetails Added Settings
