+++ app/src/main/java/com/example/eateri/MainActivity.kt
@@ -66,6 +66,7 @@ class MainActivity : AppCompatActivity(){
                 R.id.nav_myProfile,
                 R.id.nav_myOrders,
                 R.id.myPaymentFragment,
+                R.id.nav_setting,
                 R.id.nav_register,
                 R.id.loginFragment
             ), drawerLayout
