+++ app/src/main/java/com/example/eateri/ui/settings/setting.kt
@@ -1,4 +1,4 @@
-package com.example.eateri.ui
+package com.example.eateri.ui.settings
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
