+++ app/src/main/res/layout/fragment_setting.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="?attr/backgroundColor"
-    tools:context=".ui.setting">
+    tools:context=".ui.settings.setting">
 
     <!-- TODO: Update blank fragment layout -->
