+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -74,4 +74,9 @@
             android:name="position"
             app:argType="integer" />
     </fragment>
+    <fragment
+        android:id="@+id/nav_setting"
+        android:name="com.example.eateri.ui.settings.setting"
+        android:label="Setting"
+        tools:layout="@layout/fragment_setting" />
 </navigation>
