+++ app/src/main/res/values/strings.xml
@@ -27,4 +27,5 @@
     <string name="profile_title_personInfo">PERSONAL INFORMATION</string>
     <string name="btnSave">SAVE</string>
     <string name="btnCancel">CANCEL</string>
+    <string name="navSetting">Settings</string>
 </resources>
